Move mapTooltipPropsToRadixProps to its own function/file
